Dependency checking environment
Add an tox environment that does static analysis (via pylint) without
install test-requirements.txt; this ensures that the dependencies
specified in requirements.txt/setup.py are enough. Previously, the
tests could pass but the package might not build if a necessary
dependency was specified in the test environment but not in the real
package metadata.